# openspec/specs/api-contract/openapi.yaml
openapi: "3.1.0"
info:
  title: CheckMate Secrets Detection API
  version: "1.0.0"
  description: |
    CheckMate is a self-hosted secrets detection platform. This API provides
    real-time scanning, project management, exception handling, and AI-assisted
    triage for hard-coded secrets in source code, configuration, and git history.

    ## Authentication

    Two mechanisms are supported:

    - **JWT Bearer token**: Obtained via `POST /v1/auth/token`. Short-lived (15 min)
      access tokens with refresh. Used for human/interactive sessions.
    - **API Key**: Created via `POST /v1/auth/api-keys`. Scoped, long-lived keys
      prefixed `cm_`. Used for CI/headless integration. Stored as bcrypt hashes —
      shown once on creation, never retrievable again.

    Both use `Authorization: Bearer <token>` header format.

    In development mode (`CHECKMATE_DEV_MODE=true`), authentication is bypassed.

  contact:
    name: Dr. Adedayo Adetoye
    url: https://github.com/adedayo/checkmate
  license:
    name: BSD 3-Clause
    url: https://github.com/adedayo/checkmate/blob/master/LICENSE
servers:
  - url: http://localhost:17283
    description: Local development environment
    x-internal: true
  - url: "{scheme}://{host}:{port}"
    description: Self-hosted production environment
    x-internal: false
    variables:
      scheme:
        default: http
        enum: [http, https]
      host:
        default: localhost
      port:
        default: "17283"
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer authentication token supporting RFC8725 profile."
  schemas:
...
    SecretType:
      type: string
      description: Structured secret type taxonomy
      enum:
        - aws.access_key
        - aws.secret_key
        - aws.session_token
        - gcp.service_account_key
        - azure.client_secret
        - azure.sas_token
        - github.personal_access_token
        - github.fine_grained_token
        - github.app_private_key
        - gitlab.personal_access_token
        - db.password
        - db.connection_string
        - crypto.rsa_private_key
        - crypto.ec_private_key
        - crypto.pkcs12
        - crypto.pem_private_key
        - crypto.ssh_private_key
        - generic.high_entropy
        - generic.password
        - generic.api_key
        - generic.token
        - generic.jwt
    Severity:
      type: string
      enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
    Confidence:
      type: string
      enum: [CONFIRMED, HIGH, MEDIUM, LOW]
...
    VerificationStatus:
      type: string
      enum: [NOT_CHECKED, ACTIVE, REVOKED, UNKNOWN]
    PromptMode:
      type: string
      enum: [REDACTED, RAW_VALUE]
...
    Finding:
      type: object
      required:
        - id
        - ruleId
        - secretType
        - severity
        - confidence
        - file
        - line
        - column
        - secretChecksum
        - suppressed
        - verificationStatus
        - detectedAt
      properties:
        id:
          type: string
          description: >
            Deterministic stable ID: sha256(ruleId + repoUrl + file + line + column + secretChecksum).
            Same finding across different scans always has the same ID.
          example: "sha256:a1b2c3d4e5f6..."
        ruleId:
          type: string
          example: "HIGH_ENTROPY_STRING"
        secretType:
          $ref: "#/components/schemas/SecretType"
        severity:
          $ref: "#/components/schemas/Severity"
        confidence:
          $ref: "#/components/schemas/Confidence"
        repositoryUrl:
          type: string
          nullable: true
          example: "https://github.com/org/repo"
        commitSha:
          type: string
          nullable: true
          description: Git commit SHA. Null for filesystem scans.
          example: "deadbeef1234"
        branch:
          type: string
          nullable: true
          example: "main"
        file:
          type: string
          example: "src/config/database.go"
        line:
          type: integer
          example: 42
        column:
          type: integer
          example: 20
        evidenceRedacted:
          type: string
          nullable: true
          description: First 4 + last 4 characters of the detected value. Raw value never stored.
          example: "xK9$...L3m"
        secretChecksum:
          type: string
          description: sha256 of the raw secret value. Used for dedup and exception matching.
          example: "sha256:1a2b3c4d..."
        sourceContext:
          type: string
          nullable: true
          description: Surrounding code with secret position replaced by ████.
          example: "const dbPassword = \"████\"  // line 42"
        suppressed:
          type: boolean
          default: false
        exceptionId:
          type: string
          nullable: true
        verificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        aiAnnotation:
          $ref: "#/components/schemas/AIAnnotation"
          nullable: true
        detectedAt:
          type: string
          format: date-time
    AIAnnotation:
      type: object
      required: [model, provider, promptMode, fpLikelihood, summary, generatedAt]
      properties:
        model:
          type: string
          example: "gpt-4o-mini"
        provider:
          type: string
          example: "openai"
        promptMode:
          $ref: "#/components/schemas/PromptMode"
        fpLikelihood:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Estimated probability this finding is a false positive.
        summary:
          type: string
          description: 1-2 sentence plain-English assessment.
        remediationHint:
          type: string
          nullable: true
        contextClues:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
        promptTokens:
          type: integer
        completionTokens:
          type: integer
    AISettings:
      type: object
      required: [enabled, provider, model, baseUrl, defaultPromptMode]
      properties:
        enabled:
          type: boolean
        provider:
          type: string
          example: "ollama"
        model:
          type: string
          example: "llama3"
        baseUrl:
          type: string
          example: "http://localhost:11434/v1"
        apiKey:
          type: string
          nullable: true
          example: "sk-..."
        defaultPromptMode:
          $ref: "#/components/schemas/PromptMode"
...
    Webhook:
      type: object
      required: [id, url, events, createdAt]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - scan.completed
              - scan.failed
              - finding.new
              - exception.created
              - exception.expired
        createdAt:
          type: string
          format: date-time
        lastDeliveredAt:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "SCAN_NOT_FOUND"
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /v1/settings/ai:
    get:
      operationId: getAISettings
      summary: Retrieve the current AI triage configuration
      tags: [system, settings]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AISettings"
    put:
      operationId: updateAISettings
      summary: Update AI triage configuration
      tags: [system, settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AISettings"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AISettings"

  # ── Authentication ──────────────────────────────────────────────────────────
...
  /v1/findings/{findingId}:
    get:
      operationId: getFinding
      summary: Get single finding detail
      tags: [findings]
      parameters:
        - name: findingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/findings/{findingId}/triage:
    post:
      operationId: triageFinding
      summary: Request AI triage for this finding
      tags: [findings]
      parameters:
        - name: findingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Triage queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [queued]
                  findingId:
                    type: string

  # ── Exceptions ──────────────────────────────────────────────────────────────
...
  /v1/webhooks:
    get:
      operationId: listWebhooks
      summary: List registered webhooks
      tags: [webhooks]
      responses:
        "200":
          description: Webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      operationId: createWebhook
      summary: Register a webhook
      tags: [webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                secret:
                  type: string
                  description: HMAC-SHA256 signing secret for payload verification.
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
  /v1/webhooks/{webhookId}:
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook
      tags: [webhooks]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /v1/webhooks/{webhookId}/test:
    post:
      operationId: testWebhook
      summary: Send a test event to a webhook
      tags: [webhooks]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Test event sent

  # ── Workspaces ───────────────────────────────────────────────────────────────
...